Refactor dataset module directory. Unit tested.
