+++ libs/dataset/types/tests/test_attention_dataset.py
@@ -1,13 +1,9 @@
-#
-# Unit tests for extractor
-#
-
 import unittest
 
-from attention_for_histone_modification.libs.preprocessing.attention_dataset import (
+from attention_for_histone_modification.libs.dataset.types.attention_dataset import (
         AttentionDataset, AttentionDatasetConfig)
-from attention_for_histone_modification.libs.preprocessing.attention_training_example import AttentionTrainingExample
-from attention_for_histone_modification.libs.preprocessing.tests.utilities_for_tests import (
+from attention_for_histone_modification.libs.dataset.types.attention_training_example import AttentionTrainingExample
+from attention_for_histone_modification.libs.dataset.types.tests.unittest_helpers import (
         create_attention_config_by_indices, create_training_example_by_label)
 
 class TestAttentionDataset(unittest.TestCase):
